Moved DbContext to Domain
